refactor(codegen): split pto_codegen.cpp into focused compilation units
- Extract type string utilities (DataTypeToMLIR, MemorySpaceToMLIR,
  FormatTileBufTypeString, ExtractTileTypeInfo) to pto_type_utils.h/cpp
- Extract ReorderTpopChains to tpop_chain_reorder.h/cpp, unify three
  duplicated dependency-scanning blocks into ClassifyStmtForTpopChain
- Extract scalar expression visitors (arithmetic, comparison, logical,
  bitwise, cast, unary) to pto_scalar_expr_codegen.cpp
- Extract control flow visitors (for, if, while, eval, yield) to
  pto_control_flow_codegen.cpp
- Group 30+ per-function fields into FunctionState struct with Reset(),
  replacing 20+ manual .clear() lines in GenerateFunction
- Rename misleading TopoSortFunctions to SortFunctionsByRoleAndLevel
  in distributed codegen (it sorts by role/level, not topologically)

pto_codegen.cpp reduced from 2509 to 1247 lines. No behavioral changes;
all 3179 unit tests pass.

Made-with: Cursor
